Fix PiCtory configuration after Upgrade
After upgrading from Debian Stretch to Buster PiCtory has moved
from /var/www to /var/www/revpi. We need to move the configuration
files to the new location either. Otherwise the device setup will not work
anymore as default _config.rsc is empty. This also applies to the
credentials for the login to Webstatus.

Signed-off-by: Frank Pavlic <f.pavlic@kunbus.com>