Make the device matching in udev rules more robust
The current matching just checks if the compatible string is a
substring. This might lead to preoblems later if new revisions of a
device will be released. For example a RevPi Connect 4 might have the
follwing compatible "kunbus,revpi-connect4". The rule fot the current
RevPi Connect would match the old and the new RevPi Connect.

Run grep only once to get the kunbus compatible string. End grep after
the first match. Any further matches can't be processed. And the first
matching compatible should be the right one. Then match the RESULT with
the desired compatible string.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>