Add support for RevPi Core SE and Connect SE
The RevPi Core SE and Connect SE have no ethernet on the pibridge thus
we don't need to handle the ks8851 on these devices.

The Core SE needs no special rules and the Connect SE can use the same
rules as the RevPi Connect. The rule to handle the naming of the pileft
interface will just not match anything.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>