Fix missing systemd tag
Without TAG+="systemd" systemd does not create a corresponding .device
file for the device in the udev rule. Therefore the service which is
specified in ENV{SYSTEMD_WANTS} won't be executed (ENV{SYSTEMD_WANTS}
adds a dependency to the service in the device file). For more details
see systemd.device(5).

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>