Fix broken function names to repair X login
Having dashes in the function names in revpi-functions breaks Xsession
and therefore it is not possible to login via graphical target. The login
silently fails with looping back to login after the password was entered
successful.

From ~/.xsession-errors:

"/etc/X11/Xsession: 8: /usr/share/revpi/revpi-functions: Syntax error: Bad function name"

Function names in bash should not contain - (dash). From bash manual:

"A word consisting solely of letters, numbers, and underscores, and
beginning with a letter or underscore. Names are used as shell
variable and function names. Also referred to as an identifier."

Fix this by replacing the dashed with underscores.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>