Fix an upgrade issue when using apt-get upgrade
apt-get upgrade will not install new packages. The new dependency to
python3-revpi-device-info will make apt-get to hold back this package.
With the no so strong dependency with Suggests the new package
installation is avoided and this package is not held back.

The sos report will work work without the python3-revpi-device-info but
will print an error in the report. This behavior is ok for now.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>