Update control file with dependency for avahi service
Avahi can announce network services on the local network.
In order to allow other users to find a service on the network
a configuration file for this service has to be created in
/etc/avahi/services.

An avahi-service for RevolutionPi devices has been added
to the master branch in commit 3fc50e9.

Avahi-services depends on avahi-daemon and this dependency
has been added in the debian/control file.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>