Add a script to write the eeprom of the network devices
We use different network device types (phy/mac). Currently we maintain
two scripts to write the mac address into the eeprom of our network
devices. In time we will need to support more devices. The major part of
the scripts is doing the same stuff.

* checking if the mac address is "valid"
* use ethtool to write to the eeprom
* use ethtool to write the mac address

The differences are mostly in the size and name of the image. And the location
of the mac address.

Unify the scripts into one script and add wrapper scripts to provide
backwards compatibility.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>