Move set mac and eeprom image stuff to library
With the new RevPi Connect 4 we need to add set mac and eeprom for (at least)
Microchip LAN7800. With future products, there might be even more chips to add.

In order to prevent duplication of code and make it simpler to add new chips,
introduce a library that can be included in each script.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>