Wrapper script for writing eeprom and set mac
We use different network device types (phy/mac). Currently we maintain
three scripts to write the mac address into the eeprom of our network
devices.

In order to have common interface for the user add a wrapper script which takes
the interface name, type and the mac address and calls the respective
implementation.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>