Add systemd, udev and pkg-config to Build-Depends
We are now using pkg-config to determine the location for the system
units and udev related files. The configuration for pkg-config for
systemd is provided by the systemd package. The same goes for udev with
the udev package.

Add systemd, udev and pkg-config to the Build-Depends otherwise the
installation of the systemd units or the udev related files will fail.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>