fix(udev): Remove systemd tag for hdmi-disable rule
The udev rule which triggers the hdmi-disable.service on RevPi Flat matches
every device path during ACTION="add". It is also tagged with the "systemd"
tag, which will result in lots of unwanted systemd devices files. These
device files are for example checked for dependencies on EACH ssh login due
to the usage of pam_systemd.so.

From the systemd.device man page:

"
Note that this and the other udev device properties are not taken into account
unless the device is tagged with the "systemd" tag in the udev database,
because otherwise the device is not exposed as a systemd unit (see above).
"

This behaviour is not wanted here and will lead to increased ssh login times
on a RevPi Flat.

Fix this by removing the systemd tag.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>