feat(revpi-config): Control antenna use on a cm4 module
Add new option to the revpi-config command, which can enable and disable
the `external_antenna`. This function is only available on CM4 modules.
When activating, dtparam=ant2 is inserted into the config.txt, when
deactivating, this line is removed again. The CM4 module then falls
back to the internal antenna as default.

Signed-off-by: Sven Sager <s.sager@kunbus.com>