fix(revpi-config): Set correct order for systemctl for enable / disable a service file
According to the documentation, the enable / disable must be used as the
first call. Only then should the start / stop be carried out. In
general, there would also be the possibility to implement this behavior
with the flag `--now` at enable / disable.

This was noticed because a service cannot be stopped by dependencies if
it was not deactivated before.

Signed-off-by: Sven Sager <s.sager@kunbus.com>

# see man page `enable UNIT` and `disable UNIT`