feat(revpi-config): Enable perf-governor and downclock-cpu for CM4 too
The current implementation guards "perf-governor" and "downclock-cpu" with a
check if the devices runs on a CM3(+) module. Therefore these options are not
available on CM4 based devices. This is wrong and the original intentation was
to not make the features available on devices with the first generation of CM
(like PR100102, RevPi Core).

Fix this my introducing another helper function is_not_cm1() and change both guards.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>