fix(revpi-config): Fix compute module detection
The current approach on checking if the script is running on a CM1 is
erroneous. It assumes that if it is not running on a CM3 or CM4 is will run
on a CM1. This totaly ignores the fact that it could also run on a CM4S.

Fix this by explicitly check for revision code of a CM1 and use inverted value
for is_not_cm1(). Also add guard method for CM4S and re-order the existing
revision guards. Fail if rpi_revision_code() cannot determine the board
revision code and do not try to make any assumptions.

Fixes a1e27f0e789f30487afa940cd9ddf477eae7b531 (and earlier)

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>