fix(udev): Make ttyRS485 symlinking more robust on Flat
The ttyRS485 symlink for the RS485 interfaces depend on the tty names
(ttyAMA0 and ttyS0). If the device is renamed the symlink will not be
created. Use the hardware address in the DEVPATH in the udev rule. This
way the symlink is created even if the tty is renamed.

On the CM3(+) the registers are at 0x3f... and on the CM4 at 0xf2.... As
we provide the Flat with CM3(+) and CM4S we need to make the udev rules
match either.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>