fix: Remove setting the static machine-id
According to machine-id(5), the file `/etc/machine-id` should be filled
with randomized content on first boot. This is done on the RevPi through
the `firstboot.service` with the command `systemd-machine-id-setup`(8).

This is wiped with the lines that are removed in this patch, writing
part of the sha256sum into the file instead.

Suggested-by: Sven Sager <s.sager@kunbus.com>
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>