fix(revpi-factory-reset): Do not remove sshwarn
If `/run/sshwarn` exists, the default raspian image checks if the user
`pi` still has the default password `raspberry` on login. If this is
case, a warning is issued on the terminal.
There is no need to remove this file at all as `revpi-factory-reset`
generates a new password, so the warning should never appear again.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>