doc: Rewrite ks8851-set-mac(8) in scdoc syntax
Other man pages will be (re)written in scdoc syntax as well. As it
doesn't support comments at the beginning of a file (i.e. for specifying
REUSE license information) this information is recorded in a .reuse/dep5
file in the root of the repository. This avoids littering the tools/
directory with ".license" files, one for each scdoc file.
It's usually preferred to specify the REUSE license information within
the file itself, though.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>