feat(deb): Break on revpi-base-files < 0.3.0
The file "/etc/network/interfaces.d/pibridge.conf" moved from
revpi-tools to revpi-base-files in version 4.2.1 and 0.3.0 respectively.
Thus, a "Breaks: revpi-base-files (<< 0.3.0)" is necessary in
revpi-tools because, while the installation of both packages will
succeed without issues, the file that has moved will be missing from the
system if revpi-base-files is installed with a version prior to 0.3.0.
This is considered wrong as the file should be present at all times.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>