fix: Simplify bluetooth status detection
The current implementation will always return 0 (=disabled) as bluetooth
status, if the device is already registered. This is happens because the
uart is not present as device in /dev (bound by driver).

So drop it, as the approach is likely error-prone (device might requires
specials reset logic or other stuff) and usually is handled by dedicaded
helpers (eg. revpi-bluetooth).

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>