fix(deb): add dphys-swapfile to suggests
The package dphys-swapfile can be used by revpi-config to enable or
disable a swap file.
This dependency is listed under "Suggests" as revpi-config still
works without it. Listing it anywhere else would mean it could be
installed automatically when installing revpi-tools which isn't wanted
as installing dphys-swapfile automatically activates the swap file.

Enabling usage of a swap file with dphys-swapfile is useful on RevPi
systems with low RAM that are running tasks that require a lot of RAM
(i.e. running a GUI).
Using a swap file is not recommended on RevPi systems though as the eMMC
is quickly worn-out with one.

Cc: Sven Sager <s.sager@kunbus.com>
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>