feat(firstboot): Configure RevPi at first boot
Devices with HAT-EEPROM automatically find all the required information
about the product type. These can be automatically configured at the
first start of the image by running "revpi-factory-reset".

The service is waiting for a virtual revpi-sec device so that piSerial
can read the password. If the device does not appear, the boot is
delayed by 90 seconds (default) and the automatic configuration is not
performed. The device is provided by `revpi-base-files`.

By crating an empty flag file `/boot/firmware/skip-autoconfigure` the
auto-configuration process can be disabled at first boot. This can be
used for debugging or in the end of line test to deactivate the
automatic configuration. The flag file is on the FAT partition to allow
all operating systems to write this file.

Signed-off-by: Sven Sager <s.sager@kunbus.com>