fix(firstboot): Combine first boot jobs in the firstboot.service file
The file `firstboot.service` was created to perform multiple tasks at
the first start of the system. Through the `Type=oneshot` several
`ExecStart` lines can be specified, which are processed one after the
other. After performing all tasks, the unit file is reported as
started.

The creation of the ssh keys is no longer executed via
`dpkg-reconfigure`, because the `configure` in the OpenSSH package
would complete further tasks. Among other things, the reloading of the
systemd units, which would lead to errors in the `sysinit.target`, due
to dependencies.

Signed-off-by: Sven Sager <s.sager@kunbus.com>