fix(revpi-config): update NetworkManager state
NetworkManager uses rfkill to turn WLAN on and off, but it also tracks
the WLAN state separately in
/var/lib/NetworkManager/NetworkManager.state. Since the bookworm image
250801 we deploy a default NetworkManager state file to make sure
NetworkManager doesn't turn WLAN on at first boot, but this also blocks
rfkill from configuring WLAN because the NetworkManager state always
overrides the rfkill state.

When enabling or disabling WLAN with revpi-config by calling it with the
"ieee80211" argument also make sure that NetworkManager is updated. If
it is running it can simply be updated with a nmcli command. If it
isn't, update the state file directly

Note that this *doesn't* check if the WLAN country is set, so WLAN is
still enabled or disabled if no WLAN country is set.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>