feat(revpi-config): check if WLAN country is set
When a device is produced, the WLAN is turned off and no WLAN country is
set. While it is okay to turn on WLAN if no country is set, it is
recommended to set the country beforehand.

Setting the WLAN country in a user-accessible way can only be done
through Cockpit so far, which uses raspi-config for setting the WLAN
country. This sets the kernel parameter cfg80211.ieee80211_regdom in
/boot/firmware/cmdline.txt.

The kernel module responsible for the WLAN country setting is cfg80211,
and the parameter for the country is ieee80211_regdom. This is exposed
in sysfs through /sys/module/cfg80211/parameters/ieee80211_regdom.

To reliably check if the ieee80211_regdom parameter for cfg80211 is set
a few places would have to be checked for this, namely
/boot/firmware/cmdline.txt and /etc/modprobe.d/*. A user can also
temporarily set the country by giving the parameter to modprobe(8) when
loading cfg80211, in which case it isn't set anywhere on disk.

Instead, check sysfs if the parameter is set to a specific country, and
print a warning to set the country and reboot if it isn't. Enable WLAN
in any case as long as the cfg80211 module is loaded.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>