fix(firstboot): split service, fix ordering
The firstboot.service groups multiple functionality that is independent
of each other: resizing the file system, and creating SSH host keys,
both on first boot of the system.

Resizing the file system is necessary because the image that is flashed
onto the storage device doesn't use the entire available storage and
needs to be expanded. It isn't done with systemd tools like
systemd-repart and systemd-growfs because these only work on disks with
a GPT, but we use MBR because that is what all versions of the Raspberry
Pi bootloader support on the devices we support. This could change in
the future in favour of GPT when all supported devices are supported
from factory.

The SSH host keys are created because they are initially not present on
a new system but need to be created before the SSH daemon can start.

Running on first boot is achieved through the systemd boot machinery. It
presents the first-boot-complete.target which is only reached for the
first boot of the device. Ordering before this target and defining
"ConditionFirstBoot=yes" makes sure the service is run only during first
boot.

Additionally, the revpi-firstboot-resize-fs.service also orders itself
before local-fs-pre.target to make sure it resizes the file system as
early as possible in the boot process.

For clarity, firstboot.service is renamed to
revpi-firstboot-resize-fs.service.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>